// src/types/channel.ts
export interface EmbedData {
    title?: string;
    type?: EmbedTypes;
    description?: string;
    url?: string;
    timestamp?: number;
    color?: number;
    footer?: EmbedFooterData;
    image?: EmbedImageData;
    thumbnail?: EmbedThumbnailData;
    video?: EmbedVideoData;
    provider?: EmbedProviderData;
    author?: EmbedAuthorData;
    fields?: Array<EmbedFieldData>;
}

export interface EmbedFieldData {
    name: string;
    value: string;
    inline?: boolean;
}

export interface EmbedFooterData {
    text: string;
    icon_url?: string;
    proxy_icon_url?: string;
}

export type EmbedVideoData = EmbedGenericsData;
export type EmbedImageData = EmbedGenericsData;
export type EmbedThumbnailData = EmbedGenericsData;

export interface EmbedAuthorData {
    name: string;
    url?: string;
    icon_url?: string;
    proxy_icon_url?: string;
}

export enum EmbedTypes {
    Rich = "rich",
    Image = "image",
    Video = "video",
    Gifv = "gifv",
    Article = "article",
    Link = "link"
}

export interface EmbedProviderData {
    name?: string;
    url?: string;
}

export interface EmbedGenericsData {
    url: string;
    proxy_url?: string;
    height?: number;
    width?: number;
}